Move InputField from UI kit to Onboarding app
