+++ packages/ui/src/index.ts
@@ -3,8 +3,6 @@ export { Space } from './components/Space'
 export { Button, type Props as ButtonProps } from './components/Button/Button'
 export { InputBase } from './components/InputBase'
 export type { InputBaseProps } from './components/InputBase'
-export { InputField } from './components/InputField/InputField'
-export type { InputFieldProps } from './components/InputField/InputField'
 export { Heading } from './components/Heading/Heading'
 export type { HeadingProps, PossibleHeadingVariant } from './components/Heading/Heading'
 export { Text } from './components/Text/Text'
